Se crean clases de la capa de negocio y controlador, esto en base al proyecto base. Falta analizar que métodos falta realizar para las reglas de negocio en particular de Animal Care y cuales sobran de las que están. También se debe revisar la vista para ver que métodos faltan en el controlador y cuales sobran. También falta listar los datos que se tiene previsto mostrar a nivel de aplicacion contra la base de datos. Averiguar cómo encriptar la clave.

+++ AnimalCare/src/co/org/animalcare/modelo/dao/JpaDAO.java
@@ -15,7 +15,7 @@ public abstract class JpaDAO< T extends Serializable > {
 	   public void setClazz( final Class< T > clazzToSet ){
 	      this.clazz = clazzToSet;
 	   }
-	   public T findOne( final Long id ){
+	   public T findOne( final int id ){
 	      return entityManager.find( clazz, id );
 	   }
 	   @SuppressWarnings("unchecked")
